requirements.txt
PyQt6>=6.5.0
python-mpv>=1.0.0
platformdirs>=3.0.0
